+++ src/app/shared/util/chart.util.ts
@@ -14,3 +14,35 @@ export const subtractTimeseriesDatasets = (from: TimeseriesDatumModel[], dataset
     return {value: fromDatum.value - total, date: fromDatum.date};
   });
 };
+
+export const UTCDateFormat = (date: Date) => {
+  return date.toUTCString();
+};
+
+export const dailyDateFormat =(date: Date) => {
+  const dayOfWeek = (day: number): string => {
+    switch (day) {
+      case 0:
+        return 'Sun';
+      case 1:
+        return 'Mon';
+      case 2:
+        return 'Tue';
+      case 3:
+        return 'Wed';
+      case 4:
+        return 'Thu';
+      case 5:
+        return 'Fri';
+      case 6:
+        return 'Sat';
+    }
+  };
+  return dayOfWeek(date.getUTCDay()) + ' ' + (date.getUTCMonth() + 1) + '/' + date.getUTCDate();
+};
+
+export const hourlyDateFormat = (date: Date) => {
+  const minutes = date.getUTCMinutes() < 10 ? '0' + date.getUTCMinutes() : date.getUTCMinutes();
+  return (date.getUTCMonth() + 1) + '/' + date.getUTCDate() + ' ' +
+    date.getUTCHours() + ':' + minutes;
+};
